style: two ruff findings CI caught that the local check missed
RUF046 (redundant int() around round()) and SIM102 (collapsible if) in the two
new phase-5 modules. Local ruff passed but CI's did not -- the RUF046 fix is
gated behind --unsafe-fixes locally, so 'ruff check --fix' reported it as a
hidden fix rather than applying or failing on it.

+++ custom_components/tuya_ev_charger/charge_curve.py
@@ -209,7 +209,7 @@ class ChargeCurve:
             remaining -= step
             position += step
             if remaining <= 1e-9:
-                return int(round(minutes))
+                return round(minutes)
         return None
 
     def _highest_learnt_kwh(self) -> float | None:
